Move decoding to a new module and separate errors
