+++ automerge-backend/src/encoding.rs
@@ -1,67 +1,21 @@
 use core::fmt::Debug;
 use std::{
-    borrow::Cow,
-    convert::TryFrom,
     io,
     io::{Read, Write},
-    mem, str,
+    mem,
 };
 
 use automerge_protocol as amp;
 use flate2::{bufread::DeflateEncoder, Compression};
 
-use crate::{columnar::COLUMN_TYPE_DEFLATE, error::AutomergeError};
+use crate::columnar::COLUMN_TYPE_DEFLATE;
 
 pub(crate) const DEFLATE_MIN_SIZE: usize = 256;
 
-fn err(_s: &str) -> AutomergeError {
-    AutomergeError::EncodingError
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct Decoder<'a> {
-    pub offset: usize,
-    pub last_read: usize,
-    data: Cow<'a, [u8]>,
-}
-
-impl<'a> Decoder<'a> {
-    pub fn new(data: Cow<'a, [u8]>) -> Self {
-        Decoder {
-            offset: 0,
-            last_read: 0,
-            data,
-        }
-    }
-
-    pub fn read<T: Decodable + Debug>(&mut self) -> Result<T, AutomergeError> {
-        let mut buf = &self.data[self.offset..];
-        let init_len = buf.len();
-        let val = T::decode::<&[u8]>(&mut buf).ok_or(AutomergeError::EncodingError)?;
-        let delta = init_len - buf.len();
-        if delta == 0 {
-            Err(err("buffer size didnt change..."))
-        } else {
-            self.last_read = delta;
-            self.offset += delta;
-            Ok(val)
-        }
-    }
-
-    pub fn read_bytes(&mut self, index: usize) -> Result<&[u8], AutomergeError> {
-        if self.offset + index > self.data.len() {
-            Err(AutomergeError::EncodingError)
-        } else {
-            let head = &self.data[self.offset..self.offset + index];
-            self.last_read = index;
-            self.offset += index;
-            Ok(head)
-        }
-    }
-
-    pub fn done(&self) -> bool {
-        self.offset >= self.data.len()
-    }
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    #[error(transparent)]
+    Io(#[from] io::Error),
 }
 
 /// Encodes booleans by storing the count of the same value.
@@ -100,46 +54,6 @@ impl BooleanEncoder {
             self.count.encode(&mut self.buf).ok();
         }
         ColData::new(col, self.buf)
-    }
-}
-
-/// See discussion on [`BooleanEncoder`] for the format data is stored in.
-pub(crate) struct BooleanDecoder<'a> {
-    decoder: Decoder<'a>,
-    last_value: bool,
-    count: usize,
-}
-
-impl<'a> From<Cow<'a, [u8]>> for Decoder<'a> {
-    fn from(bytes: Cow<'a, [u8]>) -> Decoder<'a> {
-        Decoder::new(bytes)
-    }
-}
-
-impl<'a> From<Cow<'a, [u8]>> for BooleanDecoder<'a> {
-    fn from(bytes: Cow<'a, [u8]>) -> Self {
-        BooleanDecoder {
-            decoder: Decoder::from(bytes),
-            last_value: true,
-            count: 0,
-        }
-    }
-}
-
-// this is an endless iterator that returns false after input is exhausted
-impl<'a> Iterator for BooleanDecoder<'a> {
-    type Item = bool;
-
-    fn next(&mut self) -> Option<bool> {
-        while self.count == 0 {
-            if self.decoder.done() && self.count == 0 {
-                return Some(false);
-            }
-            self.count = self.decoder.read().unwrap_or_default();
-            self.last_value = !self.last_value;
-        }
-        self.count -= 1;
-        Some(self.last_value)
     }
 }
 
@@ -318,251 +232,6 @@ where
         V: Encodable,
     {
         val.encode(&mut self.buf).ok();
-    }
-}
-
-/// See discussion on [`RleEncoder`] for the format data is stored in.
-#[derive(Debug)]
-pub(crate) struct RleDecoder<'a, T> {
-    pub decoder: Decoder<'a>,
-    last_value: Option<T>,
-    count: isize,
-    literal: bool,
-}
-
-impl<'a, T> From<Cow<'a, [u8]>> for RleDecoder<'a, T> {
-    fn from(bytes: Cow<'a, [u8]>) -> Self {
-        RleDecoder {
-            decoder: Decoder::from(bytes),
-            last_value: None,
-            count: 0,
-            literal: false,
-        }
-    }
-}
-
-// this decoder needs to be able to send type T or 'null'
-// it is an endless iterator that will return all 'null's
-// once input is exhausted
-impl<'a, T> Iterator for RleDecoder<'a, T>
-where
-    T: Clone + Debug + Decodable,
-{
-    type Item = Option<T>;
-
-    fn next(&mut self) -> Option<Option<T>> {
-        while self.count == 0 {
-            if self.decoder.done() {
-                return Some(None);
-            }
-            match self.decoder.read::<i64>() {
-                Ok(count) if count > 0 => {
-                    // normal run
-                    self.count = count as isize;
-                    self.last_value = self.decoder.read().ok();
-                    self.literal = false;
-                }
-                Ok(count) if count < 0 => {
-                    // literal run
-                    self.count = count.abs() as isize;
-                    self.literal = true;
-                }
-                Ok(_) => {
-                    // null run
-                    // FIXME(jeffa5): handle usize > i64 here somehow
-                    self.count = self.decoder.read::<usize>().unwrap() as isize;
-                    self.last_value = None;
-                    self.literal = false;
-                }
-                Err(e) => {
-                    tracing::warn!(error=?e, "error during rle decoding");
-                    return None;
-                }
-            }
-        }
-        self.count -= 1;
-        if self.literal {
-            Some(self.decoder.read().ok())
-        } else {
-            Some(self.last_value.clone())
-        }
-    }
-}
-
-/// See discussion on [`DeltaEncoder`] for the format data is stored in.
-pub(crate) struct DeltaDecoder<'a> {
-    rle: RleDecoder<'a, i64>,
-    absolute_val: u64,
-}
-
-impl<'a> From<Cow<'a, [u8]>> for DeltaDecoder<'a> {
-    fn from(bytes: Cow<'a, [u8]>) -> Self {
-        DeltaDecoder {
-            rle: RleDecoder {
-                decoder: Decoder::from(bytes),
-                last_value: None,
-                count: 0,
-                literal: false,
-            },
-            absolute_val: 0,
-        }
-    }
-}
-
-impl<'a> Iterator for DeltaDecoder<'a> {
-    type Item = Option<u64>;
-
-    fn next(&mut self) -> Option<Option<u64>> {
-        if let Some(delta) = self.rle.next()? {
-            if delta < 0 {
-                self.absolute_val -= delta.abs() as u64;
-            } else {
-                self.absolute_val += delta as u64;
-            }
-            Some(Some(self.absolute_val))
-        } else {
-            Some(None)
-        }
-    }
-}
-
-pub(crate) trait Decodable: Sized {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read;
-}
-
-impl Decodable for u8 {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        let mut buffer = [0; 1];
-        bytes.read_exact(&mut buffer).ok()?;
-        Some(buffer[0])
-    }
-}
-
-impl Decodable for u32 {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        u64::decode::<R>(bytes).and_then(|val| Self::try_from(val).ok())
-    }
-}
-
-impl Decodable for usize {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        u64::decode::<R>(bytes).and_then(|val| Self::try_from(val).ok())
-    }
-}
-
-impl Decodable for isize {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        i64::decode::<R>(bytes).and_then(|val| Self::try_from(val).ok())
-    }
-}
-
-impl Decodable for i32 {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        i64::decode::<R>(bytes).and_then(|val| Self::try_from(val).ok())
-    }
-}
-
-impl Decodable for i64 {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        leb128::read::signed(bytes).ok()
-    }
-}
-
-impl Decodable for f64 {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        let mut buffer = [0; 8];
-        bytes.read_exact(&mut buffer).ok()?;
-        Some(Self::from_le_bytes(buffer))
-    }
-}
-
-impl Decodable for f32 {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        let mut buffer = [0; 4];
-        bytes.read_exact(&mut buffer).ok()?;
-        Some(Self::from_le_bytes(buffer))
-    }
-}
-
-impl Decodable for u64 {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        leb128::read::unsigned(bytes).ok()
-    }
-}
-
-impl Decodable for Vec<u8> {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        let len = usize::decode::<R>(bytes)?;
-        if len == 0 {
-            return Some(vec![]);
-        }
-        let mut buffer = vec![0; len];
-        bytes.read_exact(buffer.as_mut_slice()).ok()?;
-        Some(buffer)
-    }
-}
-impl Decodable for String {
-    fn decode<R>(bytes: &mut R) -> Option<String>
-    where
-        R: Read,
-    {
-        let buffer = Vec::decode(bytes)?;
-        str::from_utf8(&buffer).map(|t| t.into()).ok()
-    }
-}
-
-impl Decodable for Option<String> {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        let buffer = Vec::decode(bytes)?;
-        if buffer.is_empty() {
-            return Some(None);
-        }
-        Some(str::from_utf8(&buffer).map(|t| t.into()).ok())
-    }
-}
-
-impl Decodable for amp::ActorId {
-    fn decode<R>(bytes: &mut R) -> Option<Self>
-    where
-        R: Read,
-    {
-        let buffer = Vec::decode(bytes)?;
-        Some(buffer.into())
     }
 }
